cmd migrate  to cog
