+++ discordbot.py
@@ -43,9 +43,6 @@ class DiscordBot(commands.Bot):
             except Exception:
                 traceback.print_exc()
 
-        self.playlistApplicationService = PlaylistApplicationService()
-        self.songApplicationService = SongApplicationService()
-
     # set event
     # @commands.event
     # async def on_command_error(self, ctx, error):
@@ -55,43 +52,6 @@ class DiscordBot(commands.Bot):
     #    await ctx.send(error_msg)
 
     # set command
-
-    @commands.group(invoke_without_command=True)
-    async def add(self, ctx, arg_title, arg_url):
-        # add song to the playlist
-        self.songApplicationService.add(arg_title, arg_url)
-        # return console
-        await ctx.send("タイトル「" + arg_title + "」のプレイリストに追加したよ！")
-        # messageはエラーが出なかった場合のみにしたい。重複のときはエラーにしたい。
-        # 重複--> mapper指定での追加のときはかぶりが多く発生するので悩みどころ。
-
-    @add.command(alises=['pl', 'playlist'])
-    async def pl_add(self, ctx, arg_title):
-        # get playlist path
-        self.playlistApplicationService.create(arg_title)
-        # return console
-        await ctx.send("プレイリストを作成したよ！タイトル「" + arg_title + "」を使って曲を追加してね！")
-
-    @commands.group(invoke_without_command=True, aliases=['del'])
-    async def delete(self, ctx, arg_title, arg_url):
-        # del song to the playlist
-        self.songApplicationService.delete(arg_title, arg_url)
-        # return console
-        await ctx.send("タイトル「" + arg_title + "」のプレイリストから削除したよ！")
-
-    @delete.command(alises=['pl', 'playlist'])
-    async def pl_del(self, ctx, arg_title):
-        # get playlist path
-        self.playlistApplicationService.delete(arg_title)
-        # return console
-        await ctx.send("タイトル「" + arg_title + "」のプレイリストを削除したよ！")
-
-    @commands.command(aliases=['dl'])
-    async def download(self, ctx, arg_title):
-        # get playlist url
-        playlistUrl = self.playlistApplicationService.getDownloadUrl(arg_title)
-        # return console
-        await ctx.send("これをお使い！ " + playlistUrl)
 
 
 # botのインスタンス化及び起動処理
